requirements.txt
# Core Dependencies
streamlit==1.29.0
google-generativeai==0.3.2
python-dotenv==1.0.0

# Image Processing
Pillow==10.1.0

# File Processing
pandas==2.1.4
python-docx==1.1.0
PyPDF2==3.0.1

# Data Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Web Utilities
requests==2.31.0

# Excel Support
openpyxl==3.1.2

# Development & Utilities
streamlit-option-menu==0.3.6